requirements.txt
streamlit>=1.40.0
openai-whisper
requests>=2.31.0
yt-dlp
python-dotenv>=1.0.0
